backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-groq==0.1.0
langchain-core==0.1.23
langchain-community==0.0.20
pinecone-client==3.0.0
sentence-transformers==2.3.1
PyPDF2==3.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
groq==0.4.2
tenacity==8.2.3
